Remove "#if 0" block from ExpressionData
In commit 7c8b4fe088 (May 2021), the old expression library's
"make_var" function was gotten rid of, and a block of code that used
it was put inside an "#if 0" comment.  The block comprised the only
code inside an "else" block of a C++ conditional (not a compiler
conditional).  An actual C++ comment saying "This should never happen
now" precedes it.

The block of code is now removed since the function it called no
longer exists.

Xyce/backlog/project-backlog#6