Remove "#if 0" and "#if 1" constructs from Op code
In commit 2074f643d9 (May 2020) "mucho development on leads, power,
noise for .print" was done.

In N_UTL_Op.C, some of that work involved inserting and then
commenting out debug code.  And in N_UTL_Op.h, a block of code with a
comment saying it was probably unnecessary, but compiled fine, was
left with a "#if 1"/"#endif" around it indicating unconditional
compilation.

I am removing the commented-out debugging code in the .C file and the
unconditional conditional around the block of code in the .h file.

It may be the case that the code that had been inside the "#if 1" is
indeed unnecessary, but there are already regular comments there
pointing it out, so the conditional block is unnecessary to show it.

Xyce/backlog/project-backlog#6