Removes multiple unused ifdefs and unused class
This commit removes Xyce_INCLUDE_LIBPAPI and XYCE_HAS_MPI ifdefs
as well as the TimeBlockSynchronized class that has never been
used.  In looking at the code coverage, graciously provided by Tom,
and generally searching there is quite a bit of code in these files
that isn't being used.  This modest removal just addresses code that
is related to ifdefs being removed in support of the issue:

Xyce/backlog/project-backlog#7