Removes HAVE_SRENVIRONMENT_H ifdef
The HAVE_SRENVIRONMENT_H ifdef was for the ARA EMP environment,
which was a DTRA library.  According to Rich:

"The ARA EMP environment was a DTRA library that appears to have
been abandoned.  They haven’t done any code updates in years."

Rich has approved of the removal of the ifdef and code within the
ifdef in support of issue:

Xyce/backlog/project-backlog#7