Removes unused demangling code and fixes demangling
There is several ifdefs in the demangling code for GNU and IBM
compilers that is not used.  Therefore the strings passed to the
Xyce::demangle method are not demangled.  There is a function
in Teuchos::TypeNameTraits that can provide portable demangling
of a type id.  So, Xyce::demangle simply calls the Teuchos function.

Xyce/backlog/project-backlog#7