Fixes handling of M_PI across code base to be uniform
M_PI is not part of the C++ standard, but is often defined in the
<cmath> header.  This is definitely not true for Windows compilers,
but adding "#define _USE_MATH_DEFINES" before including the
cmath header allows M_PI to be defined.  Again, it is not guaranteed
that M_PI is defined, so a backup is defining it yourself.  This
was done in the N_UTL_Math.h as well, but it was done through a
calculation.  This would cause extra computation every time that
M_PI is used.  Changing the definition of M_PI to a number with a
set precision is more often used.

Changes have been made to N_UTL_Math.h to more cleanly handle
M_PI.  Places where <cmath> have been included have been changed
to N_UTL_Math.h to ensure that M_PI is defined uniformly across the
code.

Xyce/backlog/project-backlog#7