Returning M_PI definition to B3SOI device
Some of the BSIM3 devices have a less accurate set of constants,
which are probably used in the reference SPICE models.  So, changing
this version of M_PI is not necessary and could possible cause small
changes to the numeric behavior of the model.

I'm returning this model back to its former behavior.

Xyce/backlog/project-backlog#7