Removes #ifdef REUSE_PARAMPTR
Tom has informed me that the history of this ifdef is:

This came in in commit d68954cd by Dave Shirley in 2006.  The
BSIM SOI device used to reuse size dependent parameters if one
already existed (a technique copied from SPICE3F5's implementation)
but this is incompatible with using Xyce with analyses that change
things under the  hood --- it could be that size dependent
parameters need recomputing if other parameters have changed.
So he simply ifdef'd the thing that preserves them, and opted
instead to recompute them every time.

Xyce/backlog/project-backlog#7