Removes USE_DEFL ifdefs
These ifdefs have been guarding a block of code since 2005 and there
is no mechanism in autoconf or cmake to enable it.  According to Tom:

Most MOSFETS will use default L and W from the device options block
if L and W are not specified.  The BSIM3 and BSIM-SOI don't unless
this ifdef is set, and nothing does that.

So, these ifdefs are being removed.

Xyce/backlog/project-backlog#7