Add option for code coverage enabled cmake build
This commit starts introducing support for doing code coverage analysis
of Xyce. If an Instrumented build type is specified, then additional
compiler options are added to CMAKE_CXX_FLAGS. Only gcc is suppored with
this change. The default options are the code coverage flags
-fprofile-arcs and -ftest-coverage, but these options can be overwritten
on the command line. Running tests with an instrumented serial gcc build
of Xyce should generate code coverage information in the build
directory.

Xyce/backlog/project-backlog#84