CMake fix Mac SuperBuild LAPACK lib paths
1. Change the separator in the LAPACK_LIBRARIES to be '|' instead of ';' and
adding the argument "LIST_SEPARATOR |" to the
ExternalProject_Add(Trilnos ...) command.

The SuperBuild may find LAPACK installations that have ';' in their
path. This causes the line

'-DTPL_LAPACK_LIBRARIES=${LAPACK_LIBRARIES}'

to be interpreted through the ExternalProject_Add(Trilinos ...) call as
having spaces instead of ';' between LAPACK libraries. So CMake
interprets this as a command line option which results in an error.

Fix this by changing the separator in the LAPACK_LIBRARIES to be '|' and
adding the argument "LIST_SEPARATOR |" to the
ExternalProject_Add(Trilnos ...) command.

2. Set the option Xyce_USE_SUPERBUILD to default to OFF

NOTE: For the Mac OS Accelerate framework LAPACK_LIBRARIES has '\' as a
separator.