Have CMake regenerate dev version at build time
With development versions of Xyce, the reported version should include
the Git SHA against which the code was compiled along with the time of
compilation. Prior to this change, CMake would include the SHA and time
at *configure* time, not compilation time. That is now fixed.

I had to use a combination of `add_custom_command` and
`add_custom_target`. The `add_custom_command` runs the CMake script in
"BuildTimeStamp.cmake", which first creates a string that has the build
time and status of the repository as reported by `git describe --dirty`.
It then creates a `timestamp.h` file, included by XyceLib, that defines
a variable with that timestamp/SHA information.

To ensure the commands run every time, I added a dummy file dependency
called "MakeMeRunEveryBuild.h". That makes the build system look for the
file, and when it is not found, it runs `add_custom_command` to generate
it. Since "MakeMeRunEveryBuild.h" never exists, it forces
"BuildTimeStamp.cmake" to be run every time Xyce is built.

Finally, I made XyceLib dependent on the custom target with
`add_dependencies` to ensure the time stamp is built before Xyce.

Xyce/backlog/project-backlog#260
Xyce/code/uur-proprietary/Xyce!41