Removes TRILINOS_DEV ifdef
The TRILINOS_DEV ifdef was created to manage changes to Trilinos
interfaces between 12.12 and development.  The policy has changed
in that identified issues are checked during the configuration
process and handled using specific ifdefs.  So, the TRILINOS_DEV
ifdef is no longer needed and should be removed.

Xyce/backlog/project-backlog#7