Update Xyce CMake to use Trilinos CMake targets
Trilinos 13.5 is using modern CMake targets. This means that the
components in the Trilinos package can be directly referenced.
Previously, downstream users of Trilinos relied on CMake variables that
were populated with necessary linking information (separate variables
for things like library and component names, library paths, include
directories, etc.). With the modern target approach, most of these
properties are now inherited automatically when the target is linked.
Several of the old Trilinos CMake variables have also been deprecated.

This commit updates the CMake scripts in Xyce to link and reference the
Trilinos component targets instead of using the CMake variables. This
change breaks backwards compatibility with older versions of Trilinos.
The new minimum required version of Trilinos is set to 13.5 for CMake
builds of Xyce (note that Trilinos 13.5 is a development version that
includes multiple updates/commits. This change requires a build of
Trilinos after the commit that includes the modern CMake targets).

With moving to Trilinos 13.5, I am also refactoring checks and flags
that are used to enable support for older versions of trilinos. This
includes some of the CMake calls to test include/compile certain files
to determine if the build environment can locate and use a component
(these tests were removed if they could be replaced by checking for a
CMake target).

Xyce/backlog/project-backlog#264