Add variable to extend installer lib. search path
When cpack is making an installer it requires one
specify alternate search directories for required
libraries.  This commit adds an extra variable
Xyce_INSTALLER_LIBPATH in which one can specify
extra directories, separated by a semicoln,
where cpack will search for dependent libraries.
This is most needed when using compiler or math
libraries that are not part of the base system
os.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/212