Moves compile time check for S_ISREG to a configure time check
In the file N_UTL_CheckIfValidFile.C there is a compile time error
that is thrown if neither S_ISREG or _S_ISREG is defined.  This can
be detected at configure time instead for the user to address.
Granted, the lack of a definition for both S_ISREG and _S_ISREG is
rare nowadays.  But, Windows platforms do handle defining S_ISREG
differently than most platforms.  So, this configure time check is
mostly for that platform.

This check has been implemented for autoconf and cmake.  The code
that generates a compile time error has been removed since any issue
is going to be detected at configure time.

Xyce/backlog/project-backlog#7