Cmake support for XyceCInterface
Add the needed CMakeLists.txt file to XyceCInterface directory.
The directory Xyce/utils/XyceCInterface was never converted
over to the CMake build system.  This commit adds that target
and installs the header file and cinterface library with
Xyce as well.  Some logic in the top-level CMakeLists.txt
file had to be moved into utils/CMakeLists.txt so that the
Simulink interface can be optionally built while this one
will always be built.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/285