Fixes autoconf build with Dakota and adds cmake build with Dakota
The Boost libraries that are required by Dakota have changed, so
that has been update in the autoconf infrastructure for creating
an integrated Xyce-Dakota build.

Dakota now uses cmake, so the cmake infrastructure can be amended
to include the necessary calls for creating a Xyce-Dakota build.
Dakota does not define the newer targets, so the integration is
similar to older TPLs.

Xyce/backlog/project-backlog#304