Change order of ARCHDIR/lib search
An open source user posted to the google group today regarding some
link failures that were traced to his having two versions of Trilinos
installed --- on in /usr/local/lib installed by a package manager, the
other one he'd built according to our building guide directions and
installed elsewhere.

Xyce's configure was putting "-L$ARCHDIR/lib" at the the *END* of
LDFLAGS, even while it clearly puts "-I$ARCHDIR/include" and
"-I$ARCHDIR/include/trilinos" at the *beginning* of CPPFLAGS.

This means that our configure was set up to tell the compiler to
search for libraries in the specified ARCHDIR *last*, but search for
headers in the specified ARCHDIR *first*.

This is wrong.  And it has been this way since the very first
implementaiton of our configure.ac in 2001.  Clearly, we just don't
have many users that have Trilinos installed in some system directory
in addition to a custom build installed elsewhere, or this would have
bitten someone else before today.