Add VT0 as parameter alias for MOS1 VTO
The SPICE2 level 1 MOSFET only supported VTO as a parameter for the
zero bias threshold voltage.  Similarly, SPICE2 only had "UO" for the
surface mobility.

SPICE3 introduced VT0 and U0 as aliases (it calls them  "redundant"
parameters in its source code definition of "IOPR", and technically,
in this version UO is the redundant one and U0 the real one).

Years ago, we added hackish code to Xyce to support the UO/U0 alias,
and made loud noises at a user if they picked the one that wasn't
"recommended".

It is no longer so odd in Xyce to have more than one name for the same
parameter, and it is simple to support.

This commit adds vt0 as an exact alias of vto, and replaces the
hackish code for uo/u0 with the same kind of aliasing as is used
elsewhere in the code --- we just have two addPar calls with different
parameter names, both of which dump the value into the same variable
in the device model class.

In this commit I also add the correct documentation of the new
parameters (clearly noting they are aliases) to the reference guide
tables.  I also add a note that it is a mistake to specify both
aliases with different values, but that no warning is issued if that
is done (this is the case with all other aliased parameters in Xyce,
too).

While fixing up the reference guide, I notice that it still had a
statement that there were THREE mosfet models.  This hasn't been true
for a very, very long time.  Now it just says "multiple" models.

Xyce/backlog/project-backlog#302