Revert "Add processor-specific MKL libraries to bundle"
This reverts commit f78c45caa31c055f114628b266bee36dd7c5004e.

While "libmkl_mc3.1.dylib" was critically important for our installers
to work on some newer Macs, when I added it to this script our NEW mac
build system couldn't find it.  That's because it's running a newer
version of OneAPI Intel compilers, and they apparently removed these
processor-specific shared libraries (presumably they're all in the
core library now or something).

So we don't need to try to bundle this library, and it is likely that
the new binaries we've been generating since moving to the new build
machine will run even on the newest Macs (like Perry's, that sparked
the whole issue off in the first place).

Reverting the change, since it didn't do anything anyway.