Adds missing headers that are needed to successfully build
The nonlinear objective function classes are header-only but are
missing several headers that are necessary to build the code that
is provided in the header.  Those who currently use this class may
include those headers, so it hasn't been noticed.  However, to be
self-contained, these headers should be included.

Xyce/backlog/project-backlog#380