Refactors ROL DC optimization to use measure-like syntax
This commit refactors the ROL_DC class and moves it to the
N_AN_ROL_DC_Optimization.C file.  The objectives are described
using measure-like syntax and the arguments of the objectives are
located in .DATA statements or their GIDs are located using the
Nonlinear::objectiveFunctionData classes.  This commit eliminates
the last of the hard-coded parameters used in objective function
parsing, making the ROL DC infrastructure for performing curve-fit
optimization more flexible.

Xyce/backlog/project-backlog#380