Add required include file
N_NLS_ObjectiveFunctions.h is included by N_ANP_ROL_DC_Optimization.C,
but nowhere in the chain of headers included before it is
N_IO_OutputMgr.h actually included.

N_NLS_ObjectiveFunctions.h tries to reference a member element of the
output manager class, and when *I* try to build on FreeBSD with clang
it pukes because all that is available is a forward reference.

I don't know why this wasn't caught in our RHEL7 gcc testing just now, but
all of our other testing is down at the moment and so we're probably
missing a lot of other failures on other platforms because of this
missing include.