Remove an ancient, windows-specific kludge from BSIM4
Years ago, we were defining an M_PI macro as "2.0*atan(1)" if a system
header did not define it for us.  This was bad, because it was not
properly set up with parentheses --- which was awful if one *divided*
by M_PI   1/M_PI would become 1/2*atan(1) (or 0.25*pi) instead of
1/(2*atan(1)) (1/pi).  This error was present in the code from the
beginning commit 13969fa4 in 2000.

This went on for years leading to a superstition that something was
wrong specifically on windows (the only system that didn't have M_PI
defined in system headers), and a habit of doing funny things whenever
M_PI was used in the code (like saving it into a variable before doing
things with it, or putting extraneous parens into expressions).  The
BSIM4 has just such an expression and a big block of comments
explaining it.

In commit b57a07a in 2013 this bug was finally detected and fixed, but
we did not go back and undo all the goofy, superstitious hacks that
had been put into the code in an attempt to stave off the problems it
caused.

Since I'm working in the BSIM4 today, I am removing them now.

Note that the use of a function call and multiplication to get M_PI if
it is undefined was done away with altogether in commit 3b253df just
last month, and now it's a straight floating point constant.

Tangientially related to Xyce/backlog/project-backlog#419