Add multiplier (M=) to the VCCS device
Xyce/backlog/project-backlog#167
"Add multiplier to the G (Voltage-controlled current source (VCCS))
device"

This is an overdue compatibility feature.  Multiplier parameters now
work on VCCS devices.

While working on this, I browsed the multiplier code in the inductor,
capacitor and resistor and noticed that the multiplier variable (caled
"multiplicityFactor" was being initialized to 0.0 in the constructor,
and I felt it was more appropriate to initialize it to 1.0.   This has
no practical effect (as this variable's final initialization is in its
relevant "addPar" function call), but it still felt wrong to me.  So,
this commit also fixes that minor nit in each of those devices.