Changes "Directory node not found" developer error
The "Directory node not found" error has historically been treated as
a developer error.  However, a typo within a device definition can
also result in a "Directory node not found" error, while resolving
dependent variables.  This is not useful in helping users determine
where the unknown variable came from.

At the point that this error is thrown in the topology package, the
only information that can be obtained is the name of the fully resolved
device which is referring to this unknown variable/node.  So, it will
not be possible to point to the line and filename where this came from.
Error messages are limited to having information like:

Device XCSD:G_ABMI1 refers to unknown solution node XCSD:V_V_LFWD

This change also required modifications to the parallel directory
object so that the error behavior is the same between serial and
parallel runs.

Xyce/backlog/project-backlog#151