Add code for basic processing of "version"
This is a relatively trivial change, just making sure that all version
strings specified for the BSIM4 are appropriately converted to a valid
CMC-compliant version variable (a double) even when specified via
old-school non-compliant input (e.g. strings like "4.6.1" as is common
in older versions of BSIM4 and many model cards that use them).

I have also added a quick check of the version so that a user is
warned if asking for a version older than the oldest we support or
newer than the newest we support.

Interestingly, we have a LOT of tests in the test suite that ask for
versions other than what we actually have.

Xyce/backlog/project-backlog#419