Add some version-checking infrastructure
As we change BSIM4 versions, we're going to get some parameters that
are not supported below some version, or get dropped from some later
version.

To handle this, I'm adding some flags and accessors to the
"Descriptor" class.  When declaring a parameter with addPar in the
device traits, one can set a minimum and/or maximum version for which
the parameter is valid.

A "versionIsInRange" accessor will return true if passed a version
number that is in the range of that parameter (or if no range is
given), and false if outside the range.

Nothing uses this code yet.

I will need to have *another* checker function in the DeviceEntity
class to make use of it.  In *that*, we'll actually have to loop over
parameters already specified and check that they're valid for the
version in use.

I can't have this happen directly when parameters are processed,
because "version" is *itself* a parameter to be processed (unlike
"level," which is detected very early and set in the ModelBlock).  We
can't be sure that "version" has been processed at the time we're
processing any other parameter.  We'll have to process them all, then
run through them all and check them in a separate step.  And I'll make
it so that the device *implementing* multiple versions does the call,
so as not to waste time processing parameters for devices that don't
even *have* multiple versions.

This is an indeal approach, but is in keeping with trying to stay away
from massive infrastructure changes to support "version" at a deeper
level.

Xyce/backlog/project-backlog#419