Update Xyce cmake export for Trilinos target support
With the switch to Trilinos targets for our Trilinos
library dependency, the XyceConfig.cmake file that
the Xyce project makes needed to be updated.
Now XyceConfig.cmake pulls in the Trilinos
dependency when one does a find_package(Xyce).
The default compiler is not changed, but one
can use the variables Xyce_CC_COMPILER
and Xyce_CC_COMPILER_FLAGS to get all the
compiler details as well.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/264