Rip BSIM 4.6.1 functions into their own file
This commit segregates the version-generic/version-selecting code of
the BSIM4 from the version-specific implementations of function
"guts".

At this point, adding a new version of the BSIM4 means:

   1) create a new file containing only those functions that differ from
      the 4.6.1 version, giving them version-specific names.

   2) Add those functions as private member functions of the BSIM4
      classes (instance and model)

   3) add to the BSIM4 classes any new parameters and instance/model
      member variables needed to support the new version

   4) Modify the constructors of the instance and model to set up
      function pointers to the new methods based on setting of
      "versionDouble"

Note that at this point I have NOT finished wiring up the checks to
see if parameters that exist in one version and not another are used
where they aren't supposed to be.

There are also a handful of changes needed to support 4.7.0 that are
NOT part of the cheesy version-selection tool I've created.  These
include hacking on the noise code and solver state information, and
accessing that information in the constructors and such.  That'll come
next.

Xyce/backlog/project-backlog#419