Support out of source build and Xyce's compilers
The original CMakeLists file was developed to work
in the same directory as the verilog files.  This
generated a waring that "building in place" was not
recommended.  Now that the CMakeLists file has been
moved to an install directory it should support
building in any directory.  This just required
some directory prefixes to be added.

Also, with the update to XyceConfig.cmake for Trilinos
target support, the compiler and compiler flags used
in Xyce can be picked up by this project

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/69