Fixes "Directory node not found" error in parallel
In parallel, there is a directory that assigns voltage nodes to
processors based on connectivity to device(s), etc.  This parallel
directory may report that a voltage (solution) node is bad on a
different processor from the device that added the voltage node to
the directory.  So, the new logic that creates a more useful error
could be thwarted by this circumstance.

The bad solution nodes are now broadcast to every processor, if
any are found, and the error will be generated on the processor
that owns the device connected to it.

Xyce/backlog/project-backlog#151