Add BSIM 4.7.0 parameters and variables
This commit migrates all the addPar calls, member variable
declarations, initializers, and necessary changes to the constructors
in order to support adding the BSIM4 4.7.0 version to the collection
of models accessible via LEVEL=14 or LEVEL=54 with "version" support.

It does NOT add the BSIM4 4.7.0 model yet.

Because of the nature of the changes I
am applying to the BSIM4, it was not possible to get Garrick's work
back just by reverting commits and playing git games.  I have to go
back to the state of the repository as it was before reverting his
work and manually picking apart the pieces he had changed in the 4.6.1
model and surrounding infrastructure.

While I have added ".setMinimumVersion" attributes to the new
parameters that aren't supported prior to 4.7.0, I have not written
any of the code that actually *looks* at that minimum version and
throws warnings.

But as it stands now, the changes I have made don't break anything
except DEV_PARAMS (because there are new BSIM4 parameters), and
doesn't break any tests.  So I'm committing my changes as a checkpoint.

Xyce/backlog/project-backlog#419