Remove unused defines
When I migrated the guts of the BSIM4 4.6.1 model to a separate file,
all of the uses of some `#define`s were moved with it.  I *thought* I
had removed the now-unused stuff from N_DEV_MOSFET_B4.C, but hadn't.

It's gone now.

The downside of this move is that all that define junk has to be
duplicated in each version of BSIM4 we implement.  The upside is that
it's ONLY in the files where it's used.

Xyce/backlog/project-backlog#419