Re-establish early noise flag and correlated noise
In commit @79e2637, @garng added the ability for the BSIM4 to know it
was being called in noise analysis much earlier than it used to be
possible.  This was necessary because with correlated noise the device
allocates extra nodes.

That work got reverted in commit @1acdbf5e5 because simply replacing
the BSIM4.6.1 model with 4.7.0 broke tests --- the device is not
backward compatible.  That led to our current effort to support multiple
versions of the bsim4.

Unfortunately, now that I'm ready to reinstate Garrick's work in the
new framework, I can't just revert commits or cherry pick 79e2637 ---
that latter commit does more than just add the noise analysis
notification stuff, it also *used* it in the now-unusable BSIM4 source
code.

So this commit is basically a manual cherry-pick, taking the changes
of 79e2637 from `git log -p` and editing out the changes to the BSIM4,
then applying it as a patch file.

At any rate, the capability is back and this commit does nothing but
add it (it isn't used anywhere yet).

Garrick also added changes to support correlated noise in commit
@31608d178, but the changes to the noise infrastructure were comingled
with the changes to the BSIM4, so again I could not just cherry pick
that commit.  I have had to do the same sort of manual cherry-pick on
that one.

Xyce/backlog/project-backlog#419