Wire in BSIM4 4.7.0
This commit makes the changes necessary to add BSIM4 version 4.7.0 as
an option.

It is selected by using LEVEL=14 or LEVEL=54 and specifying a
"version" parameter either as "4.7.0" or "4.7" or "4.70".

As noted in the comments to issue 419, while CMC mandates that
"version" be a floating point number, older models do not respect that
new mandate and still have versions that can only be expressed as
strings.

Internally, we're using "version" as a double, but we allow it to be
specified EITHER as the version that CMC demands OR one that can be
converted to a CMC-style double with simple string conversion.

This breaks no tests other than DEV_PARAMS, which only breaks because
BSIM4 has new parameters it didn't have before.

Note that BSIM 4.6.1 and 4.7.0 live in completely different files, but
all implement methods defined in the BSIM4 class.

To add additional versions, create a new file for the new version and
define variants of the various version-specific functions that are
defined in the old versions.

If one finds when adding a new BSIM4 model version that one must
modify functions in the base file (N_DEV_MOSFET_B4) other than the
constructor and traits definitions, then this is a sign that one must
add additional function pointers and pointer setting just as I have
done for these two models.  The idea is to keep N_DEV_MOSFET_B4.C
containing ONLY methods that are identical between versions, and the
other files (N_DEV_MOSFET_B4p61 or N_DEV_MOSFET_B4p70) contain only
the version-specific stuff.

As noted, a cleaner design would be to have a BSIM4 base class and
derivative classes for the specific versions that overload functions,
but that would require a redesign of the way Xyce selects models based
on level.

With this change, the coding for issue 419 is complete, but there are
no really adequate tests or documentation yet.

Xyce/backlog/project-backlog#419