Add parameter version checking to BSIM4
In my last commit I added code that allows a device to check whether
parameters that have been given on the model card or instance line are
valid for the specific version of the model requested.

In this commit, I make the BSIM4 call it.

There are NO instance parameters in the BSIM4 that are
version-specific, but there *are* model parameters that are only valid
for 4.7 and later.  With this code in place, a user specifying those
parameters *and* asking for a model version older than 4.7 will be
warned that the model will ignore those parameters.

Xyce/backlog/project-backlog#419