Modify -doc/-doc_cat to handle version-dependence
Now that BSIM4 supports two versions using the same model level but
different values of "VERSION" in the model card, we needed to have
some way of flagging in the code what parameters were valid for any
given model version.  I did that by having attributes to parameters
that flag minimum or maximum versions, and testing all given
parameters to make sure they're being used at the right
version (warnings are issued if a parameter is specified that is
ignored for this version).

Now it's time to fix up our documentation to take into account this
possibility.

This commit changes the output of -doc and -doc_cat so that if (and
only if) a parameter is valid only for a range of versions, the device
parameter table entry for that parameter gets an annotation saying so.

Xyce/backlog/project-backlog#419