Fix mistake in Auger recombination
Today I discovered that the Auger recombination terms that I wrote
nearly 20 years ago had a mistake.  I had assumed that the Auger
functions worked for both scaled variables and unscaled variables, but I
didn't check this carefully.  That assumption had been true for SRH, but
not Auger.  The Auger functions (in N_DEV_MaterialSupport.C) only work
when the variables are not scaled.

I was attempted to run a 1D PDE device simulation at high injection.
Auger recombination should have caused the curve to roll over at
high current, but it did not.  This scaling mistake was the reason.

In this commit, I changed the following:

1.  When the MaterialSupport functions calcRaug, pdRaugN and pdRaugP are
    called (from the 1D and 2D devices) the variables passed into those
    functions are now unscaled, and the recombination rate that is returned
    (or its derivative) is scaled afterward.

2.  The comments around the calcRaug and related functions have been
    corrected.

3.  The Auger related functions now use material-dependent Cn, Cp
    coefficients.  Before they were using values hardwired to silicon.

This change will only matter when running at high injection.  So far
none of the PDE test cases are failing, suggesting that Auger
recombination isn't a factor in any of them.

Xyce/backlog/project-backlog#448