Fix multiplicity in bsim4 noise code
When noise was added to the BSIM 4 (and indeed to all of the MOSFETS)
the multiplicity factor was not put in correctly (rather, not at all).

When BSIM 4.7.0 was added, it copied the BSIM 4.6.1 noise code and so
it, too, is missing multiplicity.

This commit adds multiplicity to the noise code for both BSIM 4.6.1
and 4.7.0.

Xyce/backlog/bugs#8