Apply multiplicity to MOS6 noise
This is the correct fix to get multiplicity into MOS6's noise code,
but it is untested.

As it turns out, when run on the simplest MOS6 netlist I could come up
with (just copying a MOS3 netlist and changing the level number), the
MOS6 throws NaN as the input and output noise.  So there is some bug
in there unrelated to multiplicity already.  I am not fixing it, and
therefore the fix of multiplicity remains as untested as the noise
model itself.

Xyce/backlog/bugs#8
See also Xyce/backlog/bugs#10