Fix/update the Trilinos CMake "initial cache" file
The initial cache file that can be used to create a typical Trilinos build for
Xyce had an error. Specifically, the
     `set( CMAKE_POSITION_INDEPENDENT_CODE TRUE)`
line was not of the proper format. All the `set()` calls in an initial cache
file must use the `CACHE` option
(see https://cmake.org/cmake/help/latest/manual/cmake.1.html#options). That is
now fixed.

While I was at it, I updated the file to build against Trilinos develop. The
file should actually work with Trilinos 13 and above. The Xyce CMake has been
updated to only work with the current Trilinos develop, which will become
Trilinos 14, so it makes sense to update this now.

This supports:
Xyce/backlog/project-backlog#384