Fixes infinite loop when .ends is missing with no .end
If there is no .end in the netlist, then a missing .ends could
result in either a infinite loop or seg fault.  Basically, when
a new CircuitContext is created for a .SUBCKT, the context level
can be checked before and after creating the new subcircuit
context.  If the parser is not at the same level, then .ENDS
must not have been encountered.  This is being trapped for now.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/7