Move setting of interdependent parameters to processParams
The BSIM4 has a big block of code that sets default parameters that
have not been specified on the model line to defaults that depend on
other parameters (whether given or not).

This code has always lived in the constructor, but does not belong
there.  The really important reason not to put stuff in the
constructor is to make sure that things work correctly when parameters
are swept (e.g. by a .STEP line).  If they're in the constructor, then
they are not updated correctly if one of the parameters they depend on
is changed.

Stuff like this should be in processParams.

Another reason to do this is that from version to version, the number
of parameters set and the defaults that they choose sometimes
changes.  This is the case between version 4.6.1 and 4.7.0 of BSIM4,
and is even more different in 4.8.2.

I am already handling differences between these versions with special
processParams4p61_ and processParams4p70_ calls, and so moving this
stuff there, where it belongs, enables a cleaner way of maintaining
these differences without polluting the constructor that all versions
have in common.

Doing this will clear the way to a cleaner approach to implementing
4.8.2, the subject of issue Xyce/backlog/project-backlog#352

Xyce/backlog/project-backlog#452