Add some information to comments
Waaaaay back when we still had a partially functional BSIM 4.5.0, a
big block of code that sets model parameters was present in the model
class's processParams function.

But this block of code had come from the SPICE3F5 implementation's
"BSIM4temp" function, most of the rest of which was stuck into the
Instance::updateTemperature function in Xyce.

I found, when updating 4.5.0 to 4.6.1 that this ripping apart of
BSIM4temp into bits that go in Instance::updateTemperature and bits
that go into Model::processParams made the upgrade difficult, and I
put it all into Instance::updateTemperature, even though it is
illogical to be there (it is a mix of code that sets instance
variables and model variables).

At the time I put in some comments about it, but in rereading those
comments today I felt it didn't actually remind me why I had done it
or why it made sense.

I've augmented the comments.

Somewhat sorta kinda related to Xyce/backlog/project-backlog#352 in
the sense that understanding old code helps one write the new stuff.