Remove last of BSIM4 defines from common file
Some time ago, when I migrated core version-specific functions to
separate files, I removed a bunch of defines from the main BSIM4
file (which is common to all versions) and into the files that
actually used those defined macros.  At that time, I had to leave
one in place in N_DEV_MOSFET_B4.C because it was used in the model
constructor.

The code in the model constructor that used this define has since been
migrated to the version-specific processParams functions in other
files, so I'm removing the now-unused define.