Put test binaries behind the BUILD_TESTING flag
Some of the tests in the test suite make use of small binaries that link into
XyceLib. In CMake, they were excluded from the "all" target, so the tests built
them at test time. Because of the approach required to put the git SHA in the
Xyce development version stamp, XyceLib was re-linked with every test, and
sometimes the tests stepped on each other.

This commit puts all the test C++ files behind the `BUILD_TESTING` flag. If
`BUILD_TESTING=TRUE`, all the binaries are built along with all the unit tests.
Three of the binaries could be tested with the unit tests, so I added them to
the CTest setup.

If Xyce is built with `BUILD_TESTING=FALSE`, all the associated tests will
fail, because the executables won't exist.

Xyce/backlog/bugs#3